# src/main/java/alchemy/model/pets/constraints/Constraint.java
package alchemy.model.pets.constraints;

import java.util.UUID;

import alchemy.model.pets.Pet;
import jakarta.persistence.Column;
import jakarta.persistence.DiscriminatorColumn;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Inheritance;
import jakarta.persistence.InheritanceType;
import jakarta.persistence.Table;

@Entity
@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
@DiscriminatorColumn(name = "constraint_type")
@Table(name = "Constraints")
public abstract class Constraint {
	
	@Id
	@GeneratedValue(strategy = GenerationType.UUID)
    @Column(name = "id", columnDefinition = "uuid")
	private UUID id;
	
	public abstract boolean apply(Pet pet);
	public abstract String getSignature();

}