# src/main/resources/openapi/api-alchemy.yaml
openapi: 3.0.0
info:
  title: API Alchemy
  version: v1
  description: Backend API for Alchemy

tags:
  - name: Auth
    description: Authentication related content

paths:
  /auth/register:
    post:
      summary: Register a new account.
      operationId: register
      tags: 
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: Account succesfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (i.e. username already taken)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
          
  /auth/login:
    post:
      summary: Request a bearer token
      operationId: login
      tags: 
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: Authentication succeeded, token issued in the response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request (i.e. missing authentication data).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/refresh:
    post:
      summary: Request a new bearer token using an issued refresh token
      operationId: refresh
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh succeeded, token issued in the response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request (i.e. missing refresh token).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
          
  /admin/genes:
    get:
      summary: Request the list of pet genes options
      operationId: getGenes
      tags:
        - Admin
      responses:
        '200':
          description: Successfully fetched the list of pet genes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneListResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a pet gene.
      operationId: createGene
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneCreationRequest'
      responses:
        '200':
          description: Successfully created the pet genes, returned the modified list of pet genes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a pet gene.
      operationId: deleteGene
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneDeletionRequest'
      responses:
        '200':
          description: Successfully deleted the pet genes, returned the modified list of pet genes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Synchronize the pet genes by sending the expected list of genes.
      operationId: synchronizeGenes
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneSynchronizationRequest'
      responses:
        '200':
          description: Successfully synchronized the list of pet genes, returned the modified list of pet genes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
          
  /admin/moves:
    get:
      summary: Request the list of pet moves
      operationId: getMoves
      tags:
        - Admin
      responses:
        '200':
          description: Successfully fetched the list of pet moves.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoveListResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a pet move.
      operationId: createMove
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCreationRequest'
      responses:
        '200':
          description: Successfully created the pet move, returned the modified list of pet moves.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoveListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a pet move.
      operationId: deleteMove
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveDeletionRequest'
      responses:
        '200':
          description: Successfully deleted the pet moves, returned the modified list of pet moves.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoveListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Modify a specific move, it's constraints and it's components.
      operationId: updateMove
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveUpdateRequest'
      responses:
        '200':
          description: Successfully synchronized the list of pet moves, returned the modified list of pet moves.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoveListResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    AccountRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username for the account creation
          example: Gown
        password:
          type: string
          description: Hashed password
          format: password
          example: e2t3dzd5rgz898fzefz9951f9aze1f

    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token used to request a new access token
          example: VCJ9eyJhbGciOiJIUzI1NiIsInR5cCI6IkpX

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          format: uuid
          description: Bearer token.
          example: d0e6011c-0adc-4a75-97cb-92b462e27a7b
        accessExpirationTime:
          type: string
          format: date-time
          description: Expiration of the Bearer token
          example: 1985-04-12T23:20:50.52Z
        refreshToken:
          type: string
          format: uuid
          description: Refresh token.
          example: d69a8eac-2dd3-4b25-bb71-72924a59b928
        refreshExpirationTime:
          type: string
          format: date-time
          description: Expiration of the Refresh token
          example: 1985-04-12T23:20:50.52Z
            
    GeneCreationRequest:
      type: object
      properties:
        geneIdentifications:
          type: array
          items:
            $ref: '#/components/schemas/GeneIdentification'
            
    GeneDeletionRequest:
      type: object
      properties:
        geneIdentifications:
          type: array
          items:
            $ref: '#/components/schemas/GeneIdentification'
            
    GeneSynchronizationRequest:
      type: object
      properties:
        geneIdentifications:
          type: array
          items:
            $ref: '#/components/schemas/GeneIdentification'
          
    GeneListResponse:
      type: object
      properties:
        genes:
          type: array
          items:
            $ref: '#/components/schemas/Gene'
            
    MoveCreationRequest:
      type: object
      properties:
        moveIdentifications:
          $ref: '#/components/schemas/MoveIdentification'
            
    MoveDeletionRequest:
      type: object
      properties:
        moveIdentifications:
          $ref: '#/components/schemas/MoveIdentification'
            
    MoveUpdateRequest:
      type: object
      properties:
        move:
          $ref: '#/components/schemas/Move'
            
    MoveListResponse:
      type: object
      properties:
        moves:
          type: array
          items:
            $ref: '#/components/schemas/Move'
          
    MoveResponse:
      type: object
      properties:
        move:
          $ref: '#/components/schemas/Move'
        
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: ERR_AUTH-T001
        description:
          type: string
          description: Error description
          example: Malformed authentication
        message:
          type: string
          description: Error message
          example: The authentication object was malformed, and the application was unable to read it.
        
    ConfirmationResponse:
      type: object
      properties:
        code:
          type: string
          description: Confirmation error code
          example: CONF_ADM-C001
        description:
          type: string
          description: Confirmation error description
          example: Genes Synchronization
        message:
          type: string
          description: Confirmation error message.
          example: Requested synchronization causes the deletion of the following gene(s) -> Unicorn horns ; Please confirm by using the confirmation header.
        confirmationKey:
          type: string
          description: Header to use to confirm the request
          example: x-confirm-sync-genes
    
    AccountSummary:
      type: object
      properties:
        username:
          type: string
          description: Account username
          example: Gown
        hearts:
          type: integer
          description: Number of hearts given to this Account
          example: 16
        stars:
          type: integer
          description: Number of stars of this Account
          example: 3798
        messages:
          type: integer
          description: Number of unread messages
          example: 3
        familiar:
          $ref: "#/components/schemas/Pet"

#    Avatar:
#      type: object
#      properties:
        # Add the relevant parts to properly display the avatar

    Pet:
      type: object
      properties:
        name:
          type: string
          description: Name of the pet
          example: Pwiky
        stats:
          type: object
          properties:
            currentHealth:
              type: integer
              description: Current health points of the pet
              example: 35
            maxHealth:
              type: integer
              description: Max health points of the pet
              example: 50
        configuration:
          type: object
          properties:
            horns:
              $ref: "#/components/schemas/Gene"
            ears:
              $ref: "#/components/schemas/Gene"
            head:
              $ref: "#/components/schemas/Gene"
            floof:
              $ref: "#/components/schemas/Gene"
            wings:
              $ref: "#/components/schemas/Gene"
            body:
              $ref: "#/components/schemas/Gene"
            tail:
              $ref: "#/components/schemas/Gene"

    GeneIdentification:
      type: object
      properties:
        image:
          type: string
          description: Image file location
          example: /assets/body/body-fox.png
        name:
          type: string
          description: Name of the gene
          example: Unicorn horn
          
    Gene:
      type: object
      properties:
        identification:
          $ref: "#/components/schemas/GeneIdentification"
        type:
          $ref: '#/components/schemas/GeneType'
        tags:
          type: array
          items: 
            type: string
            description: Tag value associated with the Gene
            example: Horn
        constraints:
          type: array
          items:
            $ref: '#/components/schemas/Constraint'
            
    GeneType:
      type: string
      description: Type of the gene
      enum:
        - HORNS
        - EARS
        - HEAD
        - FLOOF
        - BODY
        - WINGS
        - TAIL
      example: HORNS
      
    MoveIdentification:
      type: object
      properties:
        name:
          type: string
          description: Name of the move
          example: Fireball
        tags:
          type: array
          items: 
            type: string
            description: Tag value associated with the Gene
            example: Fire
            
    Move: 
      type: object
      properties:
        identification:
          $ref: '#/components/schemas/MoveIdentification'
        constraints:
          type: array
          items:
            $ref: '#/components/schemas/Constraint'
        cooldown:
          type: integer
          description: Cooldown of the move, as a number of turns.
          example: 3
        components:
          type: array
          items:
            $ref: '#/components/schemas/MoveComponent'
            
    MoveComponent:
      type: object
      required:
        - type
      oneOf:
        - $ref: '#/components/schemas/DamageComponent'
      discriminator:
        propertyName: type
        mapping:
          DAMAGE_COMPONENT: '#/components/schemas/DamageComponent'
          
    DamageComponent:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Type of the component
        baseDamage:
          type: integer
          description: Base damage amount of the move
          example: 10
        baseBypass:
          type: integer
          description: Base bypass damage amount of the move
          example: 3
        damageType:
          $ref: '#/components/schemas/DamageType'
          
    DamageType:
      type: string
      description: Damage type of the component
      enum:
        - PHYSICAL
        - MAGICAL
      example: PHYSICAL
          
    Constraint:
      type: object
      required:
        - type
      oneOf:
        - $ref: '#/components/schemas/AttributeRequirement'
        - $ref: '#/components/schemas/AttributeRestriction'
      discriminator:
        propertyName: type
        mapping:
          ATTRIBUTE_REQUIREMENT: '#/components/schemas/AttributeRequirement'
          ATTRIBUTE_RESTRICTION: '#/components/schemas/AttributeRestriction'
          
    AttributeRequirement:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Type of the constraint
        attribute:
          $ref: '#/components/schemas/Attribute'
        threshold:
          type: integer
          description: Threshold to fit with the requirement
          example: 10
          
    AttributeRestriction:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Type of the constraint
        attribute:
          $ref: '#/components/schemas/Attribute'
        threshold:
          type: integer
          description: Threshold not to reach to fit the restriction
          example: 10
          
    Attribute:
      type: string
      description: Attribute of the constraint
      enum:
        - STRENGTH
        - CONSTITUTION
        - AGILITY
        - INTELLECT
        - WILLPOWER
        - TOUGHNESS
        - PRECISION
        - FOCUS
        - MOMENTUM
        - DEFENCE
        - ADAPTATION
        - RESOLVE
        - CLARITY
        - INSTINCT
        - MASTERY
      example: INTELLECT

    Announcement:
      type: object
      properties:
        date:
          type: string
          format: date-time
          description: Date at which the announcement was made
          example: 1985-04-12T23:20:50.52Z
        title:
          type: string
          description: Title of the announcement
          example: A Strange Discovery
        image:
          type: string
          description: Relative path of the image file
          example: announcements/MyAnnouncementImage.png
        content:
          type: string
          description: Text content of the announcement to display, in HTML
          example: <p>A very long text with markup</p>

    Item:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the Item
          example: 7
        name:
          type: string
          description: Name of the item
          example: Vial of Bottled Fire
        description:
          type: string
          description: Description of the item
          example: A magical glass vial containing the living breath of a dragon.

    Event:
      type: object
      properties:
        user:
          type: string
          description: Username concerned by the event
          example: Gown
        type:
          type: string
          description: Type of the event
          example: AnsweredTopic
        description:
          type: string
          description: Description of the event
          example: Gown entered the competition "Winter".

  responses:
    UnauthorizedError:
      description: Unauthenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'